Removed the doubles to keep the floats

+++ Cilindro/Program.cs
@@ -12,8 +12,8 @@ namespace Cilindro
             string input_raio = Console.ReadLine();
             float altura = Convert.ToSingle(input_altura);
             float raio = Convert.ToSingle(input_raio);
-            double volume = Math.PI * (Math.Pow(raio, 2)) * altura;
-            double area = 2 * Math.PI * raio * (raio + altura);
+            float volume = (float)Math.PI * ((float)Math.Pow(raio, 2)) * altura;
+            float area = 2 * (float)Math.PI * raio * (raio + altura);
             Console.WriteLine("Volume = {0}\nArea de superficie = {1}"
             , volume, area);
         }
